Create hiding-project-in-portfolio
In Squarespace, you can't save your work without pushing that work live. In Squarespace 7.1, every individual Project within a Portfolio is live if that Portfolio is live — either the entire Portfolio is disabled, or your Project is visible from the get-go. If you'd like to edit a Project within a live Portfolio but keep that Project hidden from public view, inject this header code in your Portfolio's Settings > Advanced, making sure to replace the Portfolio name and Project URL with your own.